fix(round): task-2 review fix round 1 — pad isolation, net/under-par ink, bounds, honest comments
Eight review findings addressed:

- CRITICAL: rewrite holeResultSchema's conceded-arm comment to state the truth —
  the required/unbounded stored field is a deliberate placement-rule exception (no
  honest default exists), its safety rests on no legacy conceded/picked-up event
  existing anywhere (prod checked directly: swng-rounds-prod/swng-snapshots-prod
  hold zero of either as of 2026-07-29), and the lambda-first deploy order is
  mandatory, not a preference — web-first would have an old lambda silently strip
  a new bundle's `strokes` key and write the exact poison event this schema exists
  to prevent.
- IMPORTANT 1: ScorePad had no `key`, so React reused the instance across a
  `selection` change and its `conceding` state survived a tap onto a different
  cell (the sheet has no scrim) — a conceded score could land on the wrong
  player. Fixed with `key={golferId-hole}`; regression test drives the exact
  repro and was verified RED without the fix.
- IMPORTANT 2: a conceded cell rendered no net and never inked under par, since
  Cell's `gross` only recognized `kind === "strokes"`. Fixed via the new
  `scoredStrokes` accessor; extended ScorecardGrid.test.tsx to pin both.
- IMPORTANT 3: the new request bound had no test. Added the sibling `strokes`
  arm's exact three-case template (>30/0/30) for conceded, plus fractional-count
  and bare-object rejection cases.
- IMPORTANT 4: reworded two comments that asserted things no longer true —
  whs.test.ts's affirmative WHS-rule justification (replaced with the honest
  reason: whs.ts is deleted whole later in this arc) and record.ts's "a conceded
  hole never holed out" (a conceded hole does have a stroke count; the exclusion
  is a later task's job).
- Fold-ins: added `scoredStrokes(result): number | undefined` to
  round/holeResult.ts (a pure accessor, fence-allowed like cellAt) and routed all
  six hand-rolled call sites through it; added `.int()` to the sibling `strokes`
  request arm; dropped the dead "conceded" branch from four apps/web/e2e/support.ts
  signatures; fixed a stale singlesMatch.ts comment citation in fieldTest.spec.ts.

pnpm validate green: domain 406/406, contracts 157/157, web 628/628.

+++ packages/contracts/src/commands.ts
@@ -111,7 +111,9 @@ export type AddGameRequest = z.infer<typeof addGameRequestSchema>;
 // — the check would be vacuous). `scoreResultInputSchema` below re-applies the annotation on a
 // separate binding, so callers still get the z.ZodType<HoleResult> field-shape/extra-arm guarantee.
 const scoreResultInputArms = z.discriminatedUnion("kind", [
-  z.object({ kind: z.literal("strokes"), strokes: z.number().min(1).max(30) }),
+  // .int(): a fractional stroke count is nonsense on a scorecard (task-2 fix round 1 — added
+  // alongside the `conceded` arm below, which always had it; the asymmetry read as an oversight).
+  z.object({ kind: z.literal("strokes"), strokes: z.number().int().min(1).max(30) }),
   z.object({ kind: z.literal("picked-up") }),
   // Bounded like the `strokes` arm above (task-2, spec §2d): a conceded hole now carries the
   // score the group says out loud, so a client proposing one must say a plausible number — the
